[RHTAP-1078] Moved SEB.go to common package
